refactor: move DataList, DataPermissions, and helpers inside their own folders
