+++ src/components/App.jsx
@@ -18,7 +18,7 @@ import Passphrase from 'containers/Passphrase'
 import Permission from 'components/Permissions/PermissionDetails'
 import PermissionsApplication from 'components/Permissions/AppPermissions'
 import PermissionsTab from 'components/Permissions/PermissionsTab'
-import DataPermissions from 'components/Permissions/DataPermissions'
+import DataPermissions from 'components/Permissions/DataPermissions/DataPermissions'
 import Profile from 'containers/Profile'
 import Sessions from 'containers/Sessions'
 import Sidebar from 'components/Sidebar'
