+++ src/components/Permissions/PermissionsTab.jsx
@@ -3,7 +3,7 @@ import Page from 'components/Page'
 import Tabs from 'cozy-ui/transpiled/react/Tabs'
 import Tab from 'cozy-ui/transpiled/react/Tab'
 import AppList from './AppList'
-import DataList from './DataList'
+import DataList from './DataList/DataList'
 import { routes } from 'constants/routes'
 import { useParams, useNavigate } from 'react-router-dom'
 import withAllLocales from '../../lib/withAllLocales'
